allows merging isolated splits with the same label

+++ neurolabi/gui/zstackobjectsourcefactory.h
@@ -45,7 +45,7 @@ public:
   static std::string MakeDvidGraySliceEnsembleSource(neutu::EAxis axis);
   static std::string MakeSplitObjectSource();
   static std::string MakeSplitResultSource();
-  static std::string MakeSplitResultSource(int label);
+  static std::string MakeSplitResultSource(uint64_t label);
   static std::string MakeNodeAdaptorSource();
   static std::string MakeFlyEmBoundBoxSource();
   static std::string MakeFlyEmRoiSource(const std::string &roiName);
